projects/ecommerce-marketing-funnel/ingestion/requirements.txt
pandas>=2.2.0
pyarrow>=15.0.0
google-cloud-storage>=2.16.0
google-cloud-bigquery>=3.25.0
python-dotenv>=1.0.1
